i#2626 AArch64 encoder: Update script to work on encoding groups.
Instead of parsing the xml files for each instruction, the script now
operates on the 'index-by-encoding' page of the ISA XML, which makes
the script slightly simpler and easier to add support for new encoding
groups. The script includes some logic to limit the element widths for
certain instructions, if they only support a limited range.

Refs #2626

Change-Id: I57c5ef2378d6cfd9f5f26f0265f23d730d713ab7